+++ HWInterface/SSA2Interface.cc
@@ -279,7 +279,7 @@ float SSA2Interface::calculateADCLSB(ReadoutChip* pSSA2, float theVrefValue)
 }
 
 // READ REGISTER ON CHIP:
-uint16_t SSA2Interface::ReadChipReg(Chip* pSSA2, const std::string& pRegNode)
+int32_t SSA2Interface::ReadChipReg(Chip* pSSA2, const std::string& pRegNode)
 {
     setBoard(pSSA2->getBeBoardId());
     auto                     cRegMap = pSSA2->getRegMap();
